added map security to power daemon.

+++ src/cmd/powerd/power.go
@@ -91,10 +91,12 @@ func rebootList(writer http.ResponseWriter, request *http.Request) {
      powerDaemon.DaemonLog.LogError("ERROR: Unable to unmarshal nodes to be rebooted.", error)
      
      for _, value := range nodes {
-          go func(value string) {
-               error = exec.Command("./powerCont.sh", resources[value].Address, "admn", "admn", "reboot", resources[value].Outlet).Run()
-               powerDaemon.DaemonLog.LogError("ERROR: Failed to run powerCont.sh in rebootList.", error)
-          }(value)
+          if _, ok := resources[value] ; ok {
+               go func(value string) {
+                    error = exec.Command("./powerCont.sh", resources[value].Address, "admn", "admn", "reboot", resources[value].Outlet).Run()
+                    powerDaemon.DaemonLog.LogError("ERROR: Failed to run powerCont.sh in rebootList.", error)
+               }(value)
+          }
      }
 }
 
@@ -127,10 +129,12 @@ func offList(writer http.ResponseWriter, request *http.Request) {
      powerDaemon.DaemonLog.LogError("ERROR: Unable to unmarshal nodes to be turned off.", error)
      
      for _, value := range nodes {
-          go func(value string) {
-               error = exec.Command("./powerCont.sh", resources[value].Address, "admn", "admn", "off", resources[value].Outlet).Run()
-               powerDaemon.DaemonLog.LogError("ERROR: Failed to run powerCont.sh in offList.", error)
-          }(value)
+          if _, ok := resources[value] ; ok {
+               go func(value string) {
+                    error = exec.Command("./powerCont.sh", resources[value].Address, "admn", "admn", "off", resources[value].Outlet).Run()
+                    powerDaemon.DaemonLog.LogError("ERROR: Failed to run powerCont.sh in offList.", error)
+               }(value)
+          }
      }
 }
 
@@ -165,8 +169,9 @@ func statusList(writer http.ResponseWriter, request *http.Request) {
      
      for _, value := range nodes {
           _, ok := outletStatus[value]
+          _, ok2 := resources[value]
           
-          if !ok {
+          if !ok && ok2 {
                someBytes, error = exec.Command("./powerCont.sh", resources[value].Address, "admn", "admn", "status").Output()
                powerDaemon.DaemonLog.LogError("ERROR: Failed to execute powerCont.sh and get out put in power status request.", error)
 
@@ -176,8 +181,10 @@ func statusList(writer http.ResponseWriter, request *http.Request) {
                     tmpStatusFields := strings.Split(tmpStatusLines[i], " ")
 
                     for _, value2 := range nodes {
-                         if resources[value2].Address == resources[value].Address && resources[value2].Outlet == tmpStatusFields[3] {
-                              outletStatus[value2] = tmpStatusFields[13]
+                         if _, ok3 := resources[value2] ; ok3 && ok2 {
+                              if resources[value2].Address == resources[value].Address && resources[value2].Outlet == tmpStatusFields[3] {
+                                   outletStatus[value2] = tmpStatusFields[13]
+                              }
                          }
                     }
                }
